Extract common service as `user_profile_service()`
