+++ backend/src/lib.rs
@@ -1,6 +1,7 @@
 mod error;
 mod model;
 pub mod routes;
+mod service;
 
 pub type Result<T> = std::result::Result<T, error::ServiceError>;
