+++ backend/src/routes/me.rs
@@ -1,12 +1,11 @@
 use actix_session::Session;
-use actix_web::{get, web, HttpResponse, Responder};
+use actix_web::{get, web, Responder};
 use serde::Serialize;
 use serde_json::json;
 use sqlx::PgPool;
-use tracing::info;
 use utoipa::ToSchema;
 
-use crate::{error::ServiceError, model::UserRepository, SESSION_KEY};
+use crate::{error::ServiceError, service::user_profile::user_profile_service, SESSION_KEY};
 
 #[derive(Clone, Serialize, ToSchema)]
 pub struct UserInfoResponse {
@@ -16,26 +15,19 @@ pub struct UserInfoResponse {
 
 #[utoipa::path(
     responses(
-        (status = 200, body = UserInfoResponse, description = "Successfully fetched user info"),
+        (status = 200, body = UserProfile, description = "Successfully fetched user info"),
         (status = 401, body = ErrorMessage, description = "Unauthorized"),
         (status = 500, body = ErrorMessage, description = "InternalServerError"),
     )
 )]
 #[get("/me")]
 #[tracing::instrument(skip(pool, session))]
-pub async fn me(pool: web::Data<PgPool>, session: Session) -> impl Responder {
-    me_service(pool.as_ref(), session).await
-}
-
-async fn me_service(pool: &PgPool, session: Session) -> crate::Result<impl Responder> {
+pub async fn me(pool: web::Data<PgPool>, session: Session) -> crate::Result<impl Responder> {
     // TODO: extract session validation
-    let stored_user_id = session
+    let user_name = session
         .get::<String>(SESSION_KEY)
         .map_err(|_| ServiceError::InternalServerError)?
         .ok_or(ServiceError::WrongCredential)?;
-    let user = pool.get_user_by_name(&stored_user_id).await.unwrap();
-    info!(user = ?user);
-
-    let res = UserInfoResponse { name: user.name };
-    Ok(HttpResponse::Ok().json(json!(res)))
+    let user = user_profile_service(pool.as_ref(), &user_name).await?;
+    Ok(web::Json(json!(user)))
 }
