+++ backend/src/service.rs
@@ -1,0 +1,1 @@
+pub mod user_profile;
