+++ schema.json
@@ -64,7 +64,7 @@
             "content": {
               "application/json": {
                 "schema": {
-                  "$ref": "#/components/schemas/UserInfoResponse"
+                  "$ref": "#/components/schemas/UserProfile"
                 }
               }
             }
@@ -212,9 +212,9 @@
     "/users/{name}": {
       "get": {
         "tags": [
-          "self::user_info"
+          "self::user_profile"
         ],
-        "operationId": "user_info",
+        "operationId": "user_profile",
         "parameters": [
           {
             "name": "name",
@@ -231,7 +231,7 @@
             "content": {
               "application/json": {
                 "schema": {
-                  "$ref": "#/components/schemas/UserInfoResponse"
+                  "$ref": "#/components/schemas/UserProfile"
                 }
               }
             }
@@ -340,7 +340,7 @@
           }
         }
       },
-      "UserInfoResponse": {
+      "UserProfile": {
         "type": "object",
         "required": [
           "name"
